requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
boto3-stubs==1.39.9
botocore-stubs==1.38.46
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
docx2txt==0.9
exceptiongroup==1.3.0
faiss-cpu==1.11.0.post1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
litellm==1.74.7
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
mypy-boto3-bedrock-runtime==1.39.7
narwhals==1.47.1
networkx==3.2.1
numpy==2.0.2
openai==1.97.0
opik==1.8.6
packaging==25.0
pandas==2.3.1
path==17.1.0
pathlib==1.0.1
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
protobuf==6.31.1
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.0.0
sentry-sdk==2.33.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.47.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.2
types-awscrt==0.27.4
types-s3transfer==0.13.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uuid6==2025.0.1
yarl==1.20.1
zipp==3.23.0
